ev_chatbot_assistant/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
pyjwt
python-dotenv
openai  # for step 3

